refactor: code refactor
